requirements.txt
pandas
python-dotenv
transformers
accelerate
langchain
typing_extensions
pydantic
langgraph
langchain-huggingface
langchain_community
sentence-transformers
tavily-python
pinecone
pinecone-client
requests
tqdm
pymupdf4llm
beautifulsoup4
streamlit
numpy==1.26.2
